adjustments

+++ src/main/java/luke/stardew/items/ItemTomato.java
@@ -10,7 +10,6 @@ import net.minecraft.core.util.helper.Direction;
 import net.minecraft.core.world.World;
 import net.minecraft.core.world.pos.TilePosc;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.Random;
 
@@ -39,9 +38,9 @@ public class ItemTomato extends ItemFood implements IDispensable {
 
     @Override
     public void onUseByActivator(@NotNull ItemStack selfStack, @NotNull World world, @NotNull TileEntityActivator activator, @NotNull Random random, @NotNull TilePosc blockPos, @NotNull Direction direction, double offX, double offY, double offZ) {
-        ProjectileTomato projectileTomato = new ProjectileTomato(world, (double) blockPos.x() + offX, (double) blockPos.y() + offY, (double) blockPos.z() + offZ);
+        ProjectileTomato projectileTomato = new ProjectileTomato(world, blockPos.x() + offX, blockPos.y() + offY, blockPos.z() + offZ);
 
-        projectileTomato.setHeading((double) direction.offsetX() * 0.6, direction.offsetY() == 0 ? 0.1 : (double) direction.offsetY() * 0.6, (float) direction.offsetZ() * 0.6F, 1.1F, 6.0F);
+        projectileTomato.setHeading(direction.offsetX() * 0.6F, direction.offsetY() == 0 ? 0.1F : direction.offsetY() * 0.6F, direction.offsetZ() * 0.6F, 1.1F, 6.0F);
         world.entityJoinedWorld(projectileTomato);
         --selfStack.stackSize;
     }
@@ -49,7 +48,7 @@ public class ItemTomato extends ItemFood implements IDispensable {
     @Override
     public void onDispensed(@NotNull ItemStack itemStack, @NotNull World world, @NotNull Random random, @NotNull Direction direction, double x, double y, double z) {
         ProjectileTomato projectileTomato = new ProjectileTomato(world, x, y, z);
-        projectileTomato.setHeading(direction.offsetX(), (double) direction.offsetY() + 0.1, direction.offsetZ(), 1.1F, 6.0F);
+        projectileTomato.setHeading(direction.offsetX(), direction.offsetY() + 0.1F, direction.offsetZ(), 1.1F, 6.0F);
         world.entityJoinedWorld(projectileTomato);
     }
 }
